+++ frontend/pages/home/hooks/tokenHook.tsx
@@ -1,11 +1,5 @@
-import {
-  AddingAssets,
-  AddingAssetsEnum,
-  ICRC1systemAssets,
-  TokenNetwork,
-  TokenNetworkEnum,
-  AccountDefaultEnum,
-} from "@/const";
+import { AddingAssets, AddingAssetsEnum, TokenNetwork, TokenNetworkEnum, AccountDefaultEnum } from "@/const";
+import { ICRC1systemAssets } from "@/defaultsTokens";
 import { useAppSelector } from "@redux/Store";
 import { Asset } from "@redux/models/AccountModels";
 import { Token } from "@redux/models/TokenModels";
